Fixed: Minor comparison warning
Comparing signed and unsigned integers.

+++ doomsday/plugins/dehread/src/dehreader.cpp
@@ -130,7 +130,7 @@ public:
 
     bool atEnd()
     {
-        return (pos >= patch.size() || patch.at(pos) == '\0');
+        return (size_t(pos) >= patch.size() || patch.at(pos) == '\0');
     }
 
     void advance()
